requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
cryptography==38.0.1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
gTTS==2.2.4
gunicorn==20.1.0
idna==3.4
playsound==1.2.2
psycopg2==2.9.5
PyAudio==0.2.12
pycparser==2.21
pymongo==3.12.3
pypiwin32==223
pyttsx3==2.90
pytz==2022.5
pywin32==304
requests==2.28.1
six==1.16.0
SpeechRecognition==3.8.1
sqlparse==0.2.4
tzdata==2022.6
urllib3==1.26.12
whitenoise==6.2.0
